+++ www/og-image/public/TYPEWR__.TTF
